chore: move PeopleSelectorModal into common/
